Refactoring. Moving functions into classes.
